+++ src/ContactFormExtensions.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Craft Contact Form Extensions plugin for Craft CMS 3.x.
+ * Craft Contact Form Extensions plugin for Craft CMS 4.x.
  *
  * Adds extensions to the Craft CMS contact form plugin.
  */
@@ -9,37 +9,18 @@ namespace hybridinteractive\contactformextensions;
 
 use Craft;
 use craft\base\Plugin;
-use craft\contactform\events\SendEvent;
-use craft\contactform\Mailer;
-use craft\events\RegisterUrlRulesEvent;
+use craft\contactform\events\SendEvent as CraftContactFormSendEvent;
+use craft\contactform\Mailer as CraftContactFormMailer;
 use craft\events\TemplateEvent;
 use craft\helpers\App;
 use craft\mail\Message;
-use craft\services\Plugins;
 use craft\web\twig\variables\CraftVariable;
-use craft\web\UrlManager;
 use craft\web\View;
+use hybridinteractive\contactformextensions\base\Routes;
 use hybridinteractive\contactformextensions\models\Settings;
-use hybridinteractive\contactformextensions\services\ContactFormExtensionsService as ContactFormExtensionsServiceService;
 use hybridinteractive\contactformextensions\variables\ContactFormExtensionsVariable;
 use yii\base\Event;
 
-/**
- * Craft plugins are very much like little applications in and of themselves. We’ve made
- * it as simple as we can, but the training wheels are off. A little prior knowledge is
- * going to be required to write a plugin.
- *
- * For the purposes of the plugin docs, we’re going to assume that you know PHP and SQL,
- * as well as some semi-advanced concepts like object-oriented programming and PHP namespaces.
- *
- * https://craftcms.com/docs/plugins/introduction
- *
- *
- * @property ContactFormExtensionsServiceService $contactFormExtensionsService
- * @property Settings                            $settings
- *
- * @method Settings getSettings()
- */
 class ContactFormExtensions extends Plugin
 {
     // Static Properties
@@ -53,7 +34,19 @@ class ContactFormExtensions extends Plugin
      */
     public static $plugin;
 
-    public $name;
+    public ?string $name;
+
+    // Public Properties
+    // =========================================================================
+
+    public bool $hasCpSection = true;
+    public bool $hasCpSettings = true;
+    public string $schemaVersion = '1.0.1';
+
+    // Traits
+    // =========================================================================
+
+    use Routes;
 
     // Public Methods
     // =========================================================================
@@ -68,15 +61,96 @@ class ContactFormExtensions extends Plugin
      * If you have a '/vendor/autoload.php' file, it will be loaded for you automatically;
      * you do not need to load it in your init() method.
      */
-    public function init()
+    public function init(): void
     {
         parent::init();
+
         self::$plugin = $this;
 
-        if (!Craft::$app->plugins->isPluginInstalled('contact-form') && !Craft::$app->request->getIsConsoleRequest()) {
-            Craft::$app->session->setNotice(Craft::t('contact-form-extensions', 'The Contact Form plugin is not installed or activated, Contact Form Extensions does not work without it.'));
+        $this->_registerVariable();
+        $this->_registerContactFormEventListeners();
+        $this->_registerSettings();
+        $this->_registerCraftContactFormCheck();
+
+        if (Craft::$app->getRequest()->getIsCpRequest()) {
+            $this->_registerCpRoutes();
+        }
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getCpNavItem(): ?array
+    {
+        if (!$this->settings->enableDatabase) {
+            return null;
         }
 
+        $nav = parent::getCpNavItem();
+
+        $nav['label'] = Craft::t('contact-form-extensions', 'Form Submissions');
+
+        // $nav['subnav']['submissions'] = [
+        //     'label' => Craft::t('contact-form-extensions', 'Submissions'),
+        //     'url' => 'contact-form-extensions/',
+        // ];
+
+        // if (Craft::$app->getUser()->getIsAdmin()) {
+        //     $nav['subnav']['settings'] = [
+        //         'label' => Craft::t('contact-form-extensions', 'Settings'),
+        //         'url' => 'contact-form-extensions/settings',
+        //     ];
+        // }
+
+        return $nav;
+    }
+
+    // Protected Methods
+    // =========================================================================
+
+    /**
+     * Creates and returns the model used to store the plugin’s settings.
+     *
+     * @return \craft\base\Model|null
+     */
+    protected function createSettingsModel(): ?\craft\base\Model
+    {
+        return new Settings();
+    }
+
+    /**
+     * Returns the rendered settings HTML, which will be inserted into the content
+     * block on the settings page.
+     *
+     * @throws \Twig\Error\LoaderError
+     * @throws \yii\base\Exception
+     *
+     * @return string The rendered settings HTML
+     */
+    protected function settingsHtml(): ?string
+    {
+        // Get and pre-validate the settings
+        $settings = $this->getSettings();
+        $settings->validate();
+
+        // Get the settings that are being defined by the config file
+        $overrides = Craft::$app->getConfig()->getConfigFromFile(strtolower($this->handle));
+
+        return Craft::$app->view->renderTemplate(
+            'contact-form-extensions/settings',
+            [
+                'settings'  => $this->getSettings(),
+                'overrides' => array_keys($overrides),
+            ]
+        );
+    }
+
+    // Private Methods
+    // =========================================================================
+
+    private function _registerSettings(): void
+    {
+        // Settings Template
         Event::on(View::class, View::EVENT_BEFORE_RENDER_TEMPLATE, function (TemplateEvent $e) {
             if (
                 $e->template === 'settings/plugins/_settings' &&
@@ -89,27 +163,32 @@ class ContactFormExtensions extends Plugin
                 ];
             }
         });
+    }
 
-        Event::on(UrlManager::class, UrlManager::EVENT_REGISTER_CP_URL_RULES, function (RegisterUrlRulesEvent $event) {
-            $event->rules = array_merge($event->rules, [
-                'contact-form-extensions/submissions/<submissionId:\d+>'                       => 'contact-form-extensions/submissions/show-submission',
-                'contact-form-extensions/submissions/<submissionId:\d+>/<siteHandle:{handle}>' => 'contact-form-extensions/submissions/show-submission',
-            ]);
+    private function _registerVariable(): void
+    {
+        Event::on(CraftVariable::class, CraftVariable::EVENT_INIT, function (Event $event) {
+            /** @var CraftVariable $variable */
+            $variable = $event->sender;
+            $variable->set('contactFormExtensions', ContactFormExtensionsVariable::class);
         });
+    }
 
-        Event::on(Mailer::class, Mailer::EVENT_BEFORE_SEND, function (SendEvent $e) {
+    private function _registerContactFormEventListeners(): void
+    {
+        // Capture Before Send Event from Craft Contact Form plugin
+        Event::on(CraftContactFormMailer::class, CraftContactFormMailer::EVENT_BEFORE_SEND, function (CraftContactFormSendEvent $e) {
             if ($e->isSpam) {
                 return;
-                //TODO: You could add here a submission to a spam table.
             }
 
-            // Recaptcha override
-            $recaptchaTemplateOverride = false;
-            if (is_array($e->submission->message) && array_key_exists('recaptchaTemplateOverride', $e->submission->message)) {
-                $recaptchaTemplateOverride = filter_var($e->submission->message['recaptchaTemplateOverride'], FILTER_VALIDATE_BOOLEAN);
+            // Disable Recaptcha
+            $disableRecaptcha = false;
+            if (is_array($e->submission->message) && array_key_exists('disableRecaptcha', $e->submission->message)) {
+                $disableRecaptcha = filter_var($e->submission->message['disableRecaptcha'], FILTER_VALIDATE_BOOLEAN);
             }
 
-            if ($this->settings->recaptcha && $recaptchaTemplateOverride != true) {
+            if ($this->settings->recaptcha && $disableRecaptcha != true) {
                 $recaptcha = $this->contactFormExtensionsService->getRecaptcha();
                 $captchaResponse = Craft::$app->request->getParam('g-recaptcha-response');
 
@@ -121,24 +200,24 @@ class ContactFormExtensions extends Plugin
                 }
             }
 
-            // Save to DB override
-            $saveSubmissionOverride = false;
-            if (is_array($e->submission->message) && array_key_exists('saveSubmissionOverride', $e->submission->message)) {
-                $saveSubmissionOverride = filter_var($e->submission->message['saveSubmissionOverride'], FILTER_VALIDATE_BOOLEAN);
+            // Disable Saving Submission to DB
+            $disableSaveSubmission = false;
+            if (is_array($e->submission->message) && array_key_exists('disableSaveSubmission', $e->submission->message)) {
+                $disableSaveSubmission = filter_var($e->submission->message['disableSaveSubmission'], FILTER_VALIDATE_BOOLEAN);
             }
 
             $submission = $e->submission;
-            if ($this->settings->enableDatabase && $saveSubmissionOverride != true) {
+            if ($this->settings->enableDatabase && $disableSaveSubmission != true) {
                 $this->contactFormExtensionsService->saveSubmission($submission);
             }
 
-            // Set the overridden "toEmail" setting
+            // Override toEmail setting
             if (is_array($e->submission->message) && array_key_exists('toEmail', $e->submission->message)) {
                 $email = Craft::$app->security->validateData($e->submission->message['toEmail']);
                 $e->toEmails = explode(',', $email);
             }
 
-            // Override template (Notification)
+            // Notification Template and overrides
             if ($this->settings->enableTemplateOverwrite) {
                 // First set the template mode to the Site templates
                 Craft::$app->view->setTemplateMode(View::TEMPLATE_MODE_SITE);
@@ -167,32 +246,38 @@ class ContactFormExtensions extends Plugin
             }
         });
 
-        Event::on(Mailer::class, Mailer::EVENT_AFTER_SEND, function (SendEvent $e) {
+        // Capture After Send Event from Craft Contact Form plugin
+        Event::on(CraftContactFormMailer::class, CraftContactFormMailer::EVENT_AFTER_SEND, function (CraftContactFormSendEvent $e) {
+
+            // Disable confirmation
             $disableConfirmation = false;
             if (is_array($e->submission->message) && array_key_exists('disableConfirmation', $e->submission->message)) {
                 $disableConfirmation = filter_var($e->submission->message['disableConfirmation'], FILTER_VALIDATE_BOOLEAN);
             }
 
+            // Confirmation Template and overrides
             if ($this->settings->enableConfirmationEmail && $disableConfirmation != true) {
                 // First set the template mode to the Site templates
                 Craft::$app->view->setTemplateMode(View::TEMPLATE_MODE_SITE);
 
                 // Check if template is overridden in form
                 $template = null;
-                if (is_array($e->submission->message) && array_key_exists('template', $e->submission->message)) {
-                    $template = '_emails\\'.Craft::$app->security->validateData($e->submission->message['template']);
+                if (is_array($e->submission->message) && array_key_exists('confirmationTemplate', $e->submission->message)) {
+                    $template = '_emails\\'.Craft::$app->security->validateData($e->submission->message['confirmationTemplate']);
                 } else {
                     // Render the set template
                     $template = $this->settings->confirmationTemplate;
                 }
+
                 $html = Craft::$app->view->renderTemplate(
                     $template,
                     ['submission' => $e->submission]
                 );
 
-                // Create the confirmation email
+                // Check fromEmail
                 $message = new Message();
                 $message->setTo($e->submission->fromEmail);
+
                 if (isset(App::mailSettings()->fromEmail)) {
                     $message->setFrom([Craft::parseEnv(App::mailSettings()->fromEmail) => Craft::parseEnv(App::mailSettings()->fromName)]);
                 } else {
@@ -200,14 +285,14 @@ class ContactFormExtensions extends Plugin
                 }
                 $message->setHtmlBody($html);
 
-                // Check if subject is overridden in form
-                $subject = null;
-                if (is_array($e->submission->message) && array_key_exists('subject', $e->submission->message)) {
-                    $subject = Craft::$app->security->validateData($e->submission->message['subject']);
+                // Check for subject override
+                $confirmationSubject = null;
+                if (is_array($e->submission->message) && array_key_exists('confirmationSubject', $e->submission->message)) {
+                    $confirmationSubject = Craft::$app->security->validateData($e->submission->message['confirmationSubject']);
                 } else {
-                    $subject = $this->settings->getConfirmationSubject();
+                    $confirmationSubject = $this->settings->getConfirmationSubject();
                 }
-                $message->setSubject($subject);
+                $message->setSubject($confirmationSubject);
 
                 // Send the mail
                 Craft::$app->mailer->send($message);
@@ -218,71 +303,13 @@ class ContactFormExtensions extends Plugin
                 }
             }
         });
-
-        Event::on(
-            CraftVariable::class,
-            CraftVariable::EVENT_INIT,
-            function (Event $event) {
-                /** @var CraftVariable $variable */
-                $variable = $event->sender;
-                $variable->set('contactFormExtensions', ContactFormExtensionsVariable::class);
-            }
-        );
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function getCpNavItem()
+    private function _registerCraftContactFormCheck(): void
     {
-        if (!$this->settings->enableDatabase) {
-            return;
+        // Check that Craft Contact Form plugin is installed as this plugin adds to it
+        if (!Craft::$app->plugins->isPluginInstalled('contact-form') && !Craft::$app->request->getIsConsoleRequest()) {
+            Craft::$app->session->setNotice(Craft::t('contact-form-extensions', 'The Contact Form plugin is not installed or activated, Contact Form Extensions does not work without it.'));
         }
-
-        $navItem = parent::getCpNavItem();
-
-        $navItem['label'] = Craft::t('contact-form-extensions', 'Form Submissions');
-
-        return $navItem;
-    }
-
-    // Protected Methods
-    // =========================================================================
-
-    /**
-     * Creates and returns the model used to store the plugin’s settings.
-     *
-     * @return \craft\base\Model|null
-     */
-    protected function createSettingsModel()
-    {
-        return new Settings();
-    }
-
-    /**
-     * Returns the rendered settings HTML, which will be inserted into the content
-     * block on the settings page.
-     *
-     * @throws \Twig_Error_Loader
-     * @throws \yii\base\Exception
-     *
-     * @return string The rendered settings HTML
-     */
-    protected function settingsHtml(): string
-    {
-        // Get and pre-validate the settings
-        $settings = $this->getSettings();
-        $settings->validate();
-
-        // Get the settings that are being defined by the config file
-        $overrides = Craft::$app->getConfig()->getConfigFromFile(strtolower($this->handle));
-
-        return Craft::$app->view->renderTemplate(
-            'contact-form-extensions/settings',
-            [
-                'settings'  => $this->getSettings(),
-                'overrides' => array_keys($overrides),
-            ]
-        );
     }
 }