+++ src/controllers/SubmissionsController.php
@@ -4,8 +4,8 @@ namespace hybridinteractive\contactformextensions\controllers;
 
 use craft\web\Controller;
 use hybridinteractive\contactformextensions\ContactFormExtensions;
-use hybridinteractive\contactformextensions\elements\ContactFormSubmission;
-use hybridinteractive\contactformextensions\elements\db\ContactFormSubmissionQuery;
+use hybridinteractive\contactformextensions\elements\db\SubmissionQuery;
+use hybridinteractive\contactformextensions\elements\Submission;
 
 class SubmissionsController extends Controller
 {
@@ -17,19 +17,30 @@ class SubmissionsController extends Controller
      */
     public function actionShowSubmission(string $submissionId = null, string $siteHandle = null)
     {
-        $query = new ContactFormSubmissionQuery(ContactFormSubmission::class);
+        $query = new SubmissionQuery(Submission::class);
         $query->id = $submissionId;
 
-        /* @var ContactFormSubmission $submission */
+        /* @var Submission $submission */
         $submission = $query->one();
 
-        $messageObject = ContactFormExtensions::$plugin->contactFormExtensionsService->utf8AllTheThings((array) json_decode($submission->message));
-        $variables = [
-            'submission'    => $submission,
-            'siteHandle'    => $siteHandle,
-            'messageObject' => $messageObject,
-        ];
+        if ($submission) {
+            $messageObject = ContactFormExtensions::$plugin->contactFormExtensionsService->utf8AllTheThings((array) json_decode($submission->message));
 
-        return $this->renderTemplate('contact-form-extensions/submissions/_show', $variables);
+            $variables = [
+                'submission'    => $submission,
+                'siteHandle'    => $siteHandle,
+                'messageObject' => $messageObject,
+            ];
+
+            return $this->renderTemplate('contact-form-extensions/submissions/_show', $variables);
+        } else {
+            $variables = [
+                'submission'    => null,
+                'siteHandle'    => '',
+                'messageObject' => '',
+            ];
+
+            return $this->renderTemplate('contact-form-extensions/submissions/_show', $variables);
+        }
     }
 }