+++ src/migrations/Install.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Craft Contact Form Extensions plugin for Craft CMS 3.x.
+ * Craft Contact Form Extensions plugin for Craft CMS 4.x.
  *
  * Adds extensions to the Craft CMS contact form plugin.
  */
@@ -88,7 +88,7 @@ class Install extends Migration
 
         // contactform_submissions table
         $tableSchema = Craft::$app->db->schema->getTableSchema('{{%contactform_submissions}}');
-        if ($tableSchema === null) {
+        if ($tableSchema == null) {
             $tablesCreated = true;
             $this->createTable(
                 '{{%contactform_submissions}}',